requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
face_recognition==1.3.0
numpy==1.26.4
opencv-python==4.9.0.80
supabase==2.4.3
python-multipart==0.0.9
python-dotenv==1.0.1
